refactor: extract artifact logic into dedicated module
