+++ internal/sources/httpx_test.go
@@ -14,10 +14,11 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 
+	"passive-rec/internal/artifacts"
 	"passive-rec/internal/pipeline"
 )
 
-func writeArtifactsFile(t *testing.T, outdir string, artifacts []pipeline.Artifact) {
+func writeArtifactsFile(t *testing.T, outdir string, artifacts []artifacts.Artifact) {
 	t.Helper()
 	path := filepath.Join(outdir, "artifacts.jsonl")
 	file, err := os.Create(path)
@@ -36,7 +37,7 @@ func writeArtifactsFile(t *testing.T, outdir string, artifacts []pipeline.Artifa
 
 func TestHTTPXCombinesAllLists(t *testing.T) {
 	tmp := t.TempDir()
-	writeArtifactsFile(t, tmp, []pipeline.Artifact{
+	writeArtifactsFile(t, tmp, []artifacts.Artifact{
 		{Type: "domain", Value: "example.com", Up: true},
 		{Type: "domain", Value: "sub.example.com", Up: true},
 		{Type: "route", Value: "https://app.example.com/login", Up: true},
@@ -107,7 +108,7 @@ func TestHTTPXCombinesAllLists(t *testing.T) {
 
 func TestCollectHTTPXInputsDedupesAndReportsMissing(t *testing.T) {
 	tmp := t.TempDir()
-	writeArtifactsFile(t, tmp, []pipeline.Artifact{
+	writeArtifactsFile(t, tmp, []artifacts.Artifact{
 		{Type: "domain", Value: "example.com", Up: true},
 		{Type: "domain", Value: "# comment", Up: true},
 		{Type: "route", Value: "https://assets.example.com/favicon.ico", Up: true},
@@ -311,7 +312,7 @@ func TestRunHTTPXWorkersCancellation(t *testing.T) {
 
 func TestHTTPXProcessesRouteArtifactsOnly(t *testing.T) {
 	tmp := t.TempDir()
-	writeArtifactsFile(t, tmp, []pipeline.Artifact{{Type: "route", Value: "https://app.example.com/login", Up: true}})
+	writeArtifactsFile(t, tmp, []artifacts.Artifact{{Type: "route", Value: "https://app.example.com/login", Up: true}})
 
 	originalBinFinder := httpxBinFinder
 	originalRunCmd := httpxRunCmd
@@ -377,7 +378,7 @@ func TestHTTPXProcessesRouteArtifactsOnly(t *testing.T) {
 
 func TestHTTPXNormalizesOutput(t *testing.T) {
 	inputDir := t.TempDir()
-	writeArtifactsFile(t, inputDir, []pipeline.Artifact{{Type: "route", Value: "https://app.example.com", Up: true}})
+	writeArtifactsFile(t, inputDir, []artifacts.Artifact{{Type: "route", Value: "https://app.example.com", Up: true}})
 
 	originalBinFinder := httpxBinFinder
 	originalRunCmd := httpxRunCmd
@@ -472,7 +473,7 @@ func TestHTTPXNormalizesOutput(t *testing.T) {
 
 func TestHTTPXSkipsUnresponsiveResults(t *testing.T) {
 	inputDir := t.TempDir()
-	writeArtifactsFile(t, inputDir, []pipeline.Artifact{{Type: "route", Value: "https://app.example.com", Up: true}})
+	writeArtifactsFile(t, inputDir, []artifacts.Artifact{{Type: "route", Value: "https://app.example.com", Up: true}})
 
 	originalBinFinder := httpxBinFinder
 	originalRunCmd := httpxRunCmd
@@ -515,7 +516,7 @@ func TestHTTPXSkipsUnresponsiveResults(t *testing.T) {
 
 func TestHTTPXSkipsHTMLForErrorResponses(t *testing.T) {
 	inputDir := t.TempDir()
-	writeArtifactsFile(t, inputDir, []pipeline.Artifact{{Type: "route", Value: "https://missing.example.com", Up: true}})
+	writeArtifactsFile(t, inputDir, []artifacts.Artifact{{Type: "route", Value: "https://missing.example.com", Up: true}})
 
 	originalBinFinder := httpxBinFinder
 	originalRunCmd := httpxRunCmd
@@ -554,7 +555,7 @@ func TestHTTPXSkipsHTMLForErrorResponses(t *testing.T) {
 
 func TestHTTPXIncludesRedirectDomains(t *testing.T) {
 	inputDir := t.TempDir()
-	writeArtifactsFile(t, inputDir, []pipeline.Artifact{{Type: "route", Value: "https://redirect.example.com", Up: true}})
+	writeArtifactsFile(t, inputDir, []artifacts.Artifact{{Type: "route", Value: "https://redirect.example.com", Up: true}})
 
 	originalBinFinder := httpxBinFinder
 	originalRunCmd := httpxRunCmd
@@ -592,9 +593,9 @@ func TestHTTPXIncludesRedirectDomains(t *testing.T) {
 
 func TestHTTPXBatchesLargeInputs(t *testing.T) {
 	inputDir := t.TempDir()
-	var records []pipeline.Artifact
+	var records []artifacts.Artifact
 	for i := 0; i < 5; i++ {
-		records = append(records, pipeline.Artifact{Type: "route", Value: fmt.Sprintf("https://example.com/path-%d", i), Up: true})
+		records = append(records, artifacts.Artifact{Type: "route", Value: fmt.Sprintf("https://example.com/path-%d", i), Up: true})
 	}
 	writeArtifactsFile(t, inputDir, records)
 
@@ -667,7 +668,7 @@ func TestHTTPXBatchesLargeInputs(t *testing.T) {
 
 func TestHTTPXSkipsLowPriorityRoutes(t *testing.T) {
 	tmp := t.TempDir()
-	var artifactsList []pipeline.Artifact
+	var artifactsList []artifacts.Artifact
 	for _, value := range []string{
 		"https://app.example.com/login",
 		"https://app.example.com/favicon.ico",
@@ -679,7 +680,7 @@ func TestHTTPXSkipsLowPriorityRoutes(t *testing.T) {
 		"https://app.example.com/files/THUMBS.DB",
 		"https://app.example.com/assets/raw.pgm",
 	} {
-		artifactsList = append(artifactsList, pipeline.Artifact{Type: "route", Value: value, Up: true})
+		artifactsList = append(artifactsList, artifacts.Artifact{Type: "route", Value: value, Up: true})
 	}
 	writeArtifactsFile(t, tmp, artifactsList)
